Add option to skip sdist step.
First change in a sequence to allow for customization of operational
steps. The sdist creation is current unconditional, which for some
projects makes running tests unreasonably slow. For instance, OpenStack
would prefer to just have python setup.py develop run - but we'd like
to put in the support in a flexible way that will allow people to
express the pipeline needs of their project.